Sync `guild` endpoints with latest versions

+++ src/guild/modifyGuildRole.ts
@@ -12,7 +12,7 @@ export const modifyGuildRoleSchema = z.object({
       /** bitwise value of the enabled/disabled permissions */
       permissions: z.string().min(1).nullable(),
       /** RGB color value */
-      color: z.number().nullable(),
+      color: z.number().int().nullable(),
       /** whether the role should be displayed separately in the sidebar */
       hoist: z.boolean().nullable(),
       /** the role's icon image (if the guild has the `ROLE_ICONS` feature) */
@@ -26,11 +26,19 @@ export const modifyGuildRoleSchema = z.object({
 });
 
 /**
- * Modify a guild role. Requires the `MANAGE_ROLES` permission. Returns the updated role on success. Fires a [Guild Role Update](https://discord.com/developers/docs/topics/gateway#guild-role-update) Gateway event.
+ * ### [Modify Guild Role](https://discord.com/developers/docs/resources/guild#modify-guild-role)
  *
- * *This endpoint supports the `X-Audit-Log-Reason` header.*
+ * **PATCH** `/guilds/:guild/roles/:role`
  *
- * https://discord.com/developers/docs/resources/guild#modify-guild-role
+ * Modify a guild role. Requires the `MANAGE_ROLES` permission. Returns the updated {@link Role | role} on success. Fires a Guild Role Update Gateway event.
+ *
+ * > **NOTE**
+ * >
+ * > All parameters to this endpoint are optional and nullable.
+ *
+ * > **NOTE**
+ * >
+ * > This endpoint supports the `X-Audit-Log-Reason` header.
  */
 export const modifyGuildRole: Fetcher<
   typeof modifyGuildRoleSchema,
